Remove another dependencies
