+++ bzt/engine/__init__.py
@@ -1,5 +1,5 @@
 from .names import EXEC, TAURUS_ARTIFACTS_DIR, SETTINGS
 from .engine import Engine
 from .dicts import Scenario, Configuration
-from .templates import FileLister, Singletone, SelfDiagnosable
+from .templates import Singletone
 from .modules import Provisioning, Reporter, Service, Aggregator, EngineModule, ScenarioExecutor
