+++ bzt/engine/templates.py
@@ -15,33 +15,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-from abc import abstractmethod
-
-
-class FileLister(object):
-    """
-    A mixin to get required files info from executor
-    """
-
-    @abstractmethod
-    def resource_files(self):
-        """
-        Get list of resource files
-
-        :rtype: list
-        """
-        pass
 
 
 class Singletone(object):
     pass
-
-
-class SelfDiagnosable(object):
-    @abstractmethod
-    def get_error_diagnostics(self):
-        """
-
-        :rtype: list[str]
-        """
-        pass
