+++ bzt/modules/__init__.py
@@ -19,7 +19,7 @@ import abc
 import os
 
 from bzt import ToolError
-from bzt.engine import FileLister, SelfDiagnosable, ScenarioExecutor
+from bzt.engine import ScenarioExecutor
 from bzt.modules.aggregator import ConsolidatingAggregator
 from bzt.modules.console import ExecutorWidget
 from bzt.modules.functional import FunctionalAggregator, FuncSamplesReader, LoadSamplesReader
@@ -98,7 +98,7 @@ class TransactionProvider(object):
             listener.transaction_ended(self._source, label, duration)
 
 
-class SubprocessedExecutor(ReportableExecutor, FileLister, SelfDiagnosable, TransactionProvider):
+class SubprocessedExecutor(ReportableExecutor, TransactionProvider):
     """
     Class for subprocessed executors
